Composite functions for NOAUTH handshakes
* Moved `read_handshake`, `read_handshake_skip_version`, `write_handshake`, `read_auth_method`, `write_auth_method`, as well as the associated types into `mod handshake`; reexported at the old place.

* Fixed a couple minor Clippy warnings.

* Added two new functions `negotiate_noauth_with_server` and `negotiate_noauth_with_client`, as well as the error type for them `SocksV5AuthError`.
